// frontend/src/services/api/model/user.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * The Calling API
 * Central API contract for The Calling backend and frontend service integration.
 * OpenAPI spec version: 1.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  created_at?: string;
  email?: string;
  id?: string;
  name?: string;
  role?: UserRole;
}